% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_codon.R
\name{amino_acid_plot}
\alias{amino_acid_plot}
\title{Plotting amino acid counts}
\usage{
amino_acid_plot(codon_table)
}
\arguments{
\item{codon_table}{is a dataset of codons and aminoacid sympols}
}
\value{
a column plot of aminoacid counts???
}
\description{
Plotting amino acid counts
}
