% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{buildProtein}
\alias{buildProtein}
\title{Translation}
\usage{
buildProtein(codons)
}
\arguments{
\item{codons}{Codons to be translated from the mRNA to peptides according to the codon table.}
}
\value{
List of peptides building up the protein according to the mRNA.
}
\description{
Translation
}
