% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHistogram.R
\name{createHistogram}
\alias{createHistogram}
\title{Histogram Creation}
\usage{
createHistogram(dnaSeq)
}
\arguments{
\item{dnaSeq}{is the original DNA sequence}
}
\value{
this function returns a histogram
}
\description{
Histogram Creation
}
