% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA_to_RNA.R
\name{DNA_to_RNA}
\alias{DNA_to_RNA}
\title{DNA to RNA converter}
\usage{
DNA_to_RNA(DNA_string)
}
\arguments{
\item{DNA_string}{a string containing DNA}
}
\value{
The initial DNA string now converted into RNA
}
\description{
DNA to RNA converter
}
\examples{
DNA <- 'ACTTGCA'
DNA_to_RNA(DNA)
}
