% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnaSequence.R
\name{dnaSequence}
\alias{dnaSequence}
\title{Generate a DNA sequence of specified length}
\usage{
dnaSequence(dnaSize)
}
\arguments{
\item{dnaSize}{Integer. The desired length of the DNA sequence.}
}
\value{
DNA - A character string representing a DNA sequence.
}
\description{
This function generates a random DNA sequence of a specified length.
It samples nucleotides ("A", "T", "G", "C") with replacement, meaning
bases can repeat, and concatenates them into a single string.
For example, with a sequence size of 9, the output could be "AATGCCGTA".
}
