% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCodon.R
\name{extractCodon}
\alias{extractCodon}
\title{Extract codons from a DNA coding sequence}
\usage{
extractCodon(codingSequence, start = 1)
}
\arguments{
\item{codingSequence}{A string indicating the coding DNA sequence.}

\item{start}{An integer indicating the starting position for codon extraction (default is 1).}
}
\value{
A character vector containing codons.
}
\description{
Extract codons from a DNA coding sequence
}
\examples{
# Extract codons from default coding frame
extractCodon("ATCGTC")
# Extract codons from the second nucleotide
extractCodon("TATCGTC", 2)
}
