% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{figures}
\alias{figures}
\title{count of amino acids and a plot}
\usage{
figures(amino_acid_sequence)
}
\arguments{
\item{amino_acid_sequence}{}
}
\value{
plot af amino acid count
}
\description{
count of amino acids and a plot
}
