% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sequence.R
\name{generate_sequence}
\alias{generate_sequence}
\title{Generate a random DNA Seq of specified length}
\usage{
generate_sequence(sequence_length)
}
\arguments{
\item{sequence_length}{specify desired length}
}
\value{
a random DNA seq. of desired length
}
\description{
Generate a random DNA Seq of specified length
}
\examples{
generate_sequence(25)
}
