% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codons.R
\name{get_codons}
\alias{get_codons}
\title{Get codons}
\usage{
get_codons(seq, start = 1)
}
\arguments{
\item{start}{}
}
\value{
Returns the mRNA codons
}
\description{
Get codons
}
\examples{
get_codons("ACT")
}
