% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aa.R
\name{plot_aa}
\alias{plot_aa}
\title{Plot amino acid abundance}
\usage{
plot_aa(aa_seq)
}
\arguments{
\item{aa_seq}{amino acid sequence as a string}
}
\description{
Plot amino acid abundance
}
\examples{
plot_aa("YTLAYL")
}
