% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_seq.R
\name{random_seq}
\alias{random_seq}
\title{Create a random sequence of length n}
\usage{
random_seq(n)
}
\arguments{
\item{n}{}
}
\value{
Returns random DNA sequence
}
\description{
Create a random sequence of length n
}
\examples{
random_seq(10)
}
