% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomseq.R
\name{randomseq}
\alias{randomseq}
\title{Creation of random DNA sequence}
\usage{
randomseq(seq_length)
}
\arguments{
\item{seq_length}{length of the sequence to be created}
}
\value{
a sequence of length 'seq_length' in a form of a string
}
\description{
Creation of random DNA sequence
}
