% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_transcription.R
\name{rna_transcription}
\alias{rna_transcription}
\title{rna_transcription}
\usage{
rna_transcription(dna)
}
\arguments{
\item{dna}{this is a sequence of dna made up of A,T,C,G}
}
\value{
a rna sequence with T replaced with U from dna
}
\description{
rna_transcription
}
