% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription.R
\name{transcription}
\alias{transcription}
\title{Transcription - DNA to RNA converter}
\usage{
transcription(dnaString)
}
\arguments{
\item{DNA_string}{a string containing DNA}
}
\value{
The initial DNA string is now converted into RNA
}
\description{
Transcription - DNA to RNA converter
}
\examples{
DNA <- 'ACTTGCA'
transcription(DNA)
}
