% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{translation}
\alias{translation}
\title{Translation}
\usage{
translation(codons)
}
\arguments{
\item{codons}{Codons to be translated from the mRNA to peptide according to the codon table.}
}
\value{
Peptide according to the codons mRNA.
}
\description{
Translation
}
