% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_triplicate.R
\name{triplicate}
\alias{triplicate}
\title{Creates codons from the nucleotide sequense}
\usage{
triplicate(nucleotide_seq, start = 1)
}
\arguments{
\item{nucleotide_seq}{nucleotide sequence inserted}

\item{start}{place that the function starts reading}
}
\value{
returns codons
}
\description{
Creates codons from the nucleotide sequense
}
