% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_T_to_U.R
\name{TU_convertion}
\alias{TU_convertion}
\title{Transforming T to U's}
\usage{
TU_convertion(nucleotides)
}
\arguments{
\item{nucleotides}{Insert nucleotide sequence}
}
\value{
Return a new nucleotide sequence, where T has been substituted with U
}
\description{
Transforming T to U's
}
