% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_codon.R
\name{amino_acid_plot}
\alias{amino_acid_plot}
\title{Plotting amino acid counts}
\usage{
amino_acid_plot(seq)
}
\arguments{
\item{seq}{A data set of codons and amino acid symbols}
}
\value{
A column plot of the count of unique amino acid symbols
}
\description{
Plotting amino acid counts
}
