% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_to_rna.R
\name{dna_to_rna}
\alias{dna_to_rna}
\title{DNA to RNA}
\usage{
dna_to_rna(DNA_seq)
}
\arguments{
\item{DNA_seq}{provide a DNA sequence as a string object}
}
\description{
DNA to RNA
}
\examples{
dna_to_rna('ATGCATGGA')
}
