% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription.R
\name{transcription}
\alias{transcription}
\title{Get codons}
\usage{
transcription(DNA_seq)
}
\arguments{
\item{DNA_seq}{provide a DNA sequence as a string object}
}
\description{
Get codons
}
\examples{
dna_to_rna('ATGCATGGA') -> "AUGCAUGGA"
}
