% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription.R
\name{transcription}
\alias{transcription}
\title{Transcription}
\usage{
transcription(DNA_seq)
}
\arguments{
\item{DNA_seq}{provide a DNA sequence as a string object}
}
\value{
Returns the RNA sequence
}
\description{
Transcription
}
\examples{
transcription('ATGCATGGA')
}
