% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription.R
\name{transcription}
\alias{transcription}
\title{DNA transcription to RNA}
\usage{
transcription(dna_seq, start = 1)
}
\arguments{
\item{dna_seq}{An input string representing a DNA sequence}

\item{start}{An integer specifying the starting position of the trascription. Default value set as 1.}
}
\value{
A character vector of RNA codons
}
\description{
DNA transcription to RNA
}
