% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{translation}
\alias{translation}
\title{Translation}
\usage{
translation(RNA_seq)
}
\arguments{
\item{RNA_seq}{mRNA sequence in codons}
}
\value{
Returns the amino acid sequence for their three letter codes
}
\description{
Translation
}
\examples{
translation("AUGCUA")
}
