% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{translation}
\alias{translation}
\title{translation}
\usage{
translation(codons)
}
\arguments{
\item{codons}{RNA sequence consisting of three nucleotides that functions as a unit of genomic data and codes for a specific aminoacid or signals the end of protein synthesis.}

\item{polypeptide}{singlechains of amino acids linked by peptide bonds}
}
\value{
RNA sequence given for the function to operate on and it returns the aminoacids sequence
}
\description{
translation
}
